Commit v1.0

+++ src/main/java/com/szaruga/InternetBankingApplicationDemo/service/UserService.java
@@ -1,6 +1,7 @@
 package com.szaruga.InternetBankingApplicationDemo.service;
 
 import com.szaruga.InternetBankingApplicationDemo.entity.User;
+import com.szaruga.InternetBankingApplicationDemo.exception.UserNotFoundException;
 import com.szaruga.InternetBankingApplicationDemo.jpa.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -8,6 +9,8 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Predicate;
+
+import static com.szaruga.InternetBankingApplicationDemo.constants.ApplicationConstants.USER_NOT_FOUND_WITH_ID;
 
 @Service
 public class UserService {
@@ -36,7 +39,10 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(long id){
-        userRepository.deleteById(id);
+    public void deleteUser(long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(id);
+        }else throw new UserNotFoundException(USER_NOT_FOUND_WITH_ID.getMessage() + id);
     }
 }
